+++ attoparsec-aeson/src/Data/Aeson/Internal/Text.hs
@@ -1,0 +1,1 @@
+../../../../../src/Data/Aeson/Internal/Text.hs